restructure python bindings to match src structure
